feat: CS-3158 --Enhance the logging of Data Deletion Job (#100)
Description
The Data Deletion Job is responsible for delete and notify the portal credentials that are about to expire. With the current log file it is not possible to identify if the service deleted or notified the Portal regarding any credential. It is possible to know only that the Job has run well.